Rename files and add punctuation
