+++ ios/FluentUI/Shimmer/ShimmerTokenSet.swift
@@ -5,8 +5,8 @@
 
 import UIKit
 
-/// Shimmer style can be either concealing or revealing
-/// The style affects the default shimmer alpha value and the default shimmer tint color
+/// Shimmer style can be either concealing or revealing.
+/// The style affects the default shimmer alpha value and the default shimmer tint color.
 @objc public enum MSFShimmerStyle: Int, CaseIterable {
     /// Concealing shimmer: the gradient conceals parts of the subviews as it moves leaving most parts of the subviews unblocked.
     case concealing
@@ -18,15 +18,15 @@ import UIKit
 /// Design token set for the `Shimmer` control.
 public class ShimmerTokenSet: ControlTokenSet<ShimmerTokenSet.Tokens> {
     public enum Tokens: TokenSetKey {
-        /// The alpha value of the center of the gradient in the animation if shimmer is revealing shimmer
-        /// The alpha value of the view other than the gradient if shimmer is concealing shimmer
+        /// The alpha value of the center of the gradient in the animation if shimmer is revealing shimmer.
+        /// The alpha value of the view other than the gradient if shimmer is concealing shimmer.
         case shimmerAlpha
 
-        /// Tint color of the view if shimmer is revealing shimmer
-        /// Tint color of the middle of the gradient if shimmer is concealing shimmer
+        /// Tint color of the view if shimmer is revealing shimmer.
+        /// Tint color of the middle of the gradient if shimmer is concealing shimmer.
         case tintColor
 
-        /// Tint color of the view if shimmering without a cover
+        /// Tint color of the view if shimmering without a cover.
         case viewTint
 
         ///  Color of the darkest part of the shimmer's gradient.
@@ -125,6 +125,6 @@ public class ShimmerTokenSet: ControlTokenSet<ShimmerTokenSet.Tokens> {
         }
     }
 
-    /// Determines whether the shimmer is a revealing shimmer or a concealing shimmer
+    /// Determines whether the shimmer is a revealing shimmer or a concealing shimmer.
     var style: () -> MSFShimmerStyle
 }
