+++ ios/FluentUI/Shimmer/ShimmerView.swift
@@ -3,223 +3,138 @@
 //  Licensed under the MIT License.
 //
 
-import UIKit
-import Combine
+import SwiftUI
 
-/// View that converts the subviews of a container view into a loading state with the "shimmering" effect
-@objc(MSFShimmerView)
-open class ShimmerView: UIView, TokenizedControlInternal {
+/// Properties that can be used to customize the appearance of the `Shimmer`.
+@objc public protocol MSFShimmerState {
 
-    /// Optional synchronizer to sync multiple shimmer views
-    @objc open weak var animationSynchronizer: AnimationSynchronizerProtocol?
+    /// Determines whether the shimmer is a revealing shimmer or a concealing shimmer.
+    var style: MSFShimmerStyle { get set }
 
-    open override var intrinsicContentSize: CGSize { return bounds.size }
+    /// Determines whether the view itself is shimmered or an added cover on top is shimmered.
+    var shouldAddShimmeringCover: Bool { get set }
 
-    open override func layoutSubviews() {
-        super.layoutSubviews()
+    /// Whether to use the height of the view (if the view is a label), else default to token value.
+    var usesTextHeightForLabels: Bool { get set }
+}
 
-        updateViewCoverLayers()
-
-        let viewToCover = containerView ?? self
-
-        shimmeringLayer.frame = CGRect(x: -viewToCover.frame.width,
-                                       y: 0.0,
-                                       width: viewToCover.bounds.width + 2 * viewToCover.frame.width,
-                                       height: viewToCover.frame.height)
-
-        updateShimmeringLayer()
-        updateShimmeringAnimation()
-    }
-
-    /// Create a shimmer view
-    /// - Parameter containerView: view to convert layout into a shimmer -- each of containerView's first-level subviews will be mirrored
-    /// - Parameter excludedViews: subviews of `containerView` to exclude from shimmer
-    /// - Parameter animationSynchronizer: optional synchronizer to sync multiple shimmer views
-    /// - Parameter shimmerStyle: determines whether the shimmer is a revealing shimmer or a concealing shimmer
-    /// - Parameter shimmersLeafViews: True to enable shimmers to auto-adjust to font height for a UILabel -- this will more accurately reflect the text in the label rect rather than using the bounding box.
-    /// - Parameter usesTextHeightForLabels: Determines whether we shimmer the top level subviews, or the leaf nodes of the view hierarchy. If false, we use default height of 11.0
-    @objc public init(containerView: UIView? = nil,
-                      excludedViews: [UIView] = [],
-                      animationSynchronizer: AnimationSynchronizerProtocol? = nil,
-                      shimmerStyle: MSFShimmerStyle = .revealing,
-                      shimmersLeafViews: Bool = false,
-                      usesTextHeightForLabels: Bool = false) {
-        self.style = shimmerStyle
-
-        self.containerView = containerView
-        self.excludedViews = excludedViews
-        self.animationSynchronizer = animationSynchronizer
-        self.shimmersLeafViews = shimmersLeafViews
-        self.usesTextHeightForLabels = usesTextHeightForLabels
-        super.init(frame: CGRect(origin: .zero, size: containerView?.bounds.size ?? .zero))
-
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(syncAnimation),
-                                               name: UIAccessibility.reduceMotionStatusDidChangeNotification,
-                                               object: nil)
-
-        // Update appearance whenever `tokenSet` changes.
-        tokenSetSink = tokenSet.sinkChanges { [weak self] in
-            self?.updateShimmeringAnimation()
-        }
-    }
-
-    required public init?(coder: NSCoder) {
-        preconditionFailure("init(coder:) has not been implemented")
-    }
-
-    /// Manaully sync with the synchronizer
-    @objc public func syncAnimation() {
-        updateShimmeringAnimation()
-    }
-
-    // MARK: - TokenizedControl
+/// View Modifier that adds a "shimmering" effect to any view.
+public struct ShimmerView: ViewModifier, TokenizedControlView {
     public typealias TokenSetKeyType = ShimmerTokenSet.Tokens
-    public lazy var tokenSet: ShimmerTokenSet = .init(style: { self.style })
+    @ObservedObject public var tokenSet: ShimmerTokenSet
 
-    var tokenSetSink: AnyCancellable?
-
-    /// Style to draw the control.
-    public let style: MSFShimmerStyle
-
-    /// Update the frame of each layer covering views in the containerView
-    func updateViewCoverLayers() {
-        let viewToCover = containerView ?? self
-
-        viewCoverLayers.forEach { $0.removeFromSuperlayer() }
-
-        let subviews: Set<UIView> = {
-            if shimmersLeafViews {
-                var leaves = [UIView]()
-                searchLeaves(in: viewToCover, output: &leaves)
-                return Set(leaves).subtracting(Set(excludedViews))
-            } else {
-                return Set(viewToCover.subviews).subtracting(Set(excludedViews))
+    public func body(content: Content) -> some View {
+        let distance = (contentSize.width + tokenSet[.shimmerWidth].float) / cos(tokenSet[.shimmerAngle].float)
+        let duration = CFTimeInterval(distance / tokenSet[.shimmerSpeed].float) + tokenSet[.shimmerDelay].float
+        content
+            .onSizeChange { newSize in
+                contentSize = newSize
             }
-        }()
-
-        viewCoverLayers = subviews.filter({ !$0.isHidden && !($0 is ShimmerView) }).map { subview in
-            let coverLayer = CALayer()
-
-            let shouldApplyLabelCornerRadius = subview is UILabel && tokenSet[.labelCornerRadius].float >= 0
-            coverLayer.cornerRadius = shouldApplyLabelCornerRadius ? tokenSet[.labelCornerRadius].float : tokenSet[.cornerRadius].float
-            coverLayer.backgroundColor = UIColor(dynamicColor: tokenSet[.tintColor].dynamicColor).cgColor
-
-            var coverFrame = viewToCover.convert(subview.bounds, from: subview)
-            if let label = subview as? UILabel {
-                let viewLabelHeight: CGFloat? = {
-                    if usesTextHeightForLabels {
-                        return label.font.deviceLineHeight
-                    } else {
-                        return tokenSet[.labelHeight].float
-                    }
-                }()
-
-                if let viewLabelHeight = viewLabelHeight {
-                    let delta = coverFrame.height - viewLabelHeight
-                    coverFrame.size.height = viewLabelHeight
-                    coverFrame.origin.y += delta / 2
+            .modifier(AnimatedMask(tokenSet: tokenSet,
+                                   state: state,
+                                   isLabel: isLabel,
+                                   phase: phase,
+                                  contentSize: contentSize)
+                .animation(Animation.linear(duration: duration)
+                    .delay(tokenSet[.shimmerDelay].float)
+                    .repeatForever(autoreverses: false)
+            ))
+            .onAppear {
+                if !UIAccessibility.isReduceMotionEnabled {
+                    phase = .init(1.0 + (tokenSet[.shimmerWidth].float / contentSize.width))
                 }
             }
-            coverLayer.frame = coverFrame
-
-            return coverLayer
-        }
-
-        viewCoverLayers.forEach { layer.addSublayer($0) }
+            /// RTL languages require shimmer in the respective direction.
+            .flipsForRightToLeftLayoutDirection(true)
+            .matchedGeometryEffect(id: UUID(), in: self.animationId)
     }
 
-    /// Update the gradient layer that animates to provide the shimmer effect (also updates the animation)
-    func updateShimmeringLayer() {
-        let light = UIColor.white.withAlphaComponent(tokenSet[.shimmerAlpha].float).cgColor
-        let dark = UIColor(dynamicColor: tokenSet[.darkGradient].dynamicColor).cgColor
-        shimmeringLayer.colors = self.style == .concealing ? [light, dark, light] : [dark, light, dark]
+    /// An animatable modifier to interpolate between `phase` values.
+    struct AnimatedMask: AnimatableModifier {
+        var tokenSet: ShimmerTokenSet
+        var state: MSFShimmerState
+        var isLabel: Bool
+        var phase: CGFloat
+        var contentSize: CGSize
 
-        let isRTL = effectiveUserInterfaceLayoutDirection == .rightToLeft
-        let startPoint = CGPoint(x: 0.0, y: 0.5)
-        let endPoint = CGPoint(x: 1.0, y: 0.5 - tan(tokenSet[.shimmerAngle].float * (isRTL ? -1 : 1)))
-        if isRTL {
-            shimmeringLayer.startPoint = endPoint
-            shimmeringLayer.endPoint = startPoint
-        } else {
-            shimmeringLayer.startPoint = startPoint
-            shimmeringLayer.endPoint = endPoint
+        var animatableData: CGFloat {
+            get { phase }
+            set { phase = newValue }
         }
 
-        let widthPercentage = Float(tokenSet[.shimmerWidth].float / shimmeringLayer.frame.width)
-        let locationStart = NSNumber(value: 0.5 - widthPercentage / 2)
-        let locationMiddle = NSNumber(value: 0.5)
-        let locationEnd = NSNumber(value: 0.5 + widthPercentage / 2)
+        func body(content: Content) -> some View {
+            let gradientMask = GradientMask(tokenSet: tokenSet,
+                                            state: state,
+                                            contentSize: contentSize,
+                                            phase: phase)
 
-        let locations = [locationStart, locationMiddle, locationEnd]
-        shimmeringLayer.locations = isRTL ? locations.reversed() : locations
-
-        layer.mask = shimmeringLayer
-
-        updateShimmeringAnimation()
-    }
-
-    /// Update the shimmer animation
-    func updateShimmeringAnimation() {
-        shimmeringLayer.removeAnimation(forKey: "shimmering")
-
-        // For usability/accessibility reasons, the animation is not added if the user
-        // has the "reduce motion" enabled on the device.
-        guard !UIAccessibility.isReduceMotionEnabled else {
-            return
-        }
-
-        if let animationSynchronizer = animationSynchronizer, animationSynchronizer.referenceLayer == nil {
-            animationSynchronizer.referenceLayer = shimmeringLayer
-        }
-
-        let animation = CABasicAnimation(keyPath: "locations")
-
-        let widthPercentage = Float(tokenSet[.shimmerWidth].float / shimmeringLayer.frame.width)
-
-        let fromLocationStart = NSNumber(value: 0.0)
-        let fromLocationMiddle = NSNumber(value: widthPercentage / 2.0)
-        let fromLocationEnd = NSNumber(value: widthPercentage)
-
-        let toLocationStart = NSNumber(value: 1.0 - widthPercentage)
-        let toLocationMiddle = NSNumber(value: 1.0 - (widthPercentage / 2.0))
-        let toLocationEnd = NSNumber(value: 1.0)
-
-        // Do not flip values from / to for RTL. These are already relative to the layout direction.
-        animation.fromValue = [fromLocationStart, fromLocationMiddle, fromLocationEnd]
-        animation.toValue = [toLocationStart, toLocationMiddle, toLocationEnd]
-
-        let distance = (frame.width + tokenSet[.shimmerWidth].float) / cos(tokenSet[.shimmerAngle].float)
-        animation.duration = CFTimeInterval(distance / tokenSet[.shimmerSpeed].float)
-        animation.fillMode = .forwards
-
-        // Add animation (use a group to add a delay between animations)
-        let animationGroup = CAAnimationGroup()
-        animationGroup.animations = [animation]
-        animationGroup.duration = animation.duration + tokenSet[.shimmerDelay].float
-        animationGroup.repeatCount = .infinity
-        animationGroup.timeOffset = animationSynchronizer?.timeOffset(for: shimmeringLayer) ?? 0
-        shimmeringLayer.add(animationGroup, forKey: "shimmering")
-    }
-
-    /// Layers covering the subviews of the container
-    var viewCoverLayers = [CALayer]()
-
-    /// Layer that slides to provide the "shimmer" effect
-    var shimmeringLayer = CAGradientLayer()
-
-    private func searchLeaves(in view: UIView, output: inout [UIView]) {
-        for v in view.subviews {
-            if v.subviews.isEmpty && v != self {
-                output.append(v)
-            } else if !v.isHidden {
-                searchLeaves(in: v, output: &output)
+            if state.shouldAddShimmeringCover {
+                ZStack {
+                    content
+                    HStack {
+                        RoundedRectangle(cornerRadius: isLabel && tokenSet[.labelCornerRadius].float >= 0 ? tokenSet[.labelCornerRadius].float : tokenSet[.cornerRadius].float)
+                            .foregroundColor(Color.init(dynamicColor: tokenSet[.tintColor].dynamicColor))
+                            .frame(width: contentSize.width,
+                                   height: !isLabel || state.usesTextHeightForLabels ? contentSize.height : tokenSet[.labelHeight].float)
+                            .mask(gradientMask)
+                    }
+                }
+            } else {
+                content
+                    .mask(gradientMask)
             }
         }
     }
 
-    private var shimmersLeafViews: Bool
-    private var usesTextHeightForLabels: Bool
-    private var excludedViews: [UIView]
-    private weak var containerView: UIView?
+    /// A slanted, animatable gradient between light (transparent) and dark (opaque) parts of shimmer  to use as mask.
+    /// The `phase` parameter shifts the gradient, moving the shimmer band.
+    struct GradientMask: View {
+        var tokenSet: ShimmerTokenSet
+        var state: MSFShimmerState
+        var contentSize: CGSize
+        var phase: CGFloat
+
+        var body: some View {
+            let light = Color.white.opacity(self.tokenSet[.shimmerAlpha].float)
+            let dark = Color(colorValue: self.tokenSet[.darkGradient].dynamicColor.light)
+
+            let widthPercentage = tokenSet[.shimmerWidth].float / contentSize.width
+            LinearGradient(gradient: Gradient(stops: [
+                .init(color: state.style == .concealing ? light : dark, location: phase - widthPercentage),
+                .init(color: state.style == .concealing ? dark : light, location: phase - widthPercentage / 2.0),
+                .init(color: state.style == .concealing ? light : dark, location: phase)]),
+                           startPoint: .topLeading,
+                           endPoint: .bottomTrailing)
+        }
+    }
+
+    @Environment(\.fluentTheme) var fluentTheme: FluentTheme
+    @ObservedObject var state: MSFShimmerStateImpl
+
+    /// When displaying one or more shimmers, this ID will synchronize the animations.
+    let animationId: Namespace.ID
+    /// Determines whether content to shimmer is a label.
+    let isLabel: Bool
+
+    @State private var phase: CGFloat = 0
+    @State private var contentSize: CGSize = CGSize()
+}
+
+class MSFShimmerStateImpl: ControlState, MSFShimmerState {
+    @Published var style: MSFShimmerStyle
+    @Published var shouldAddShimmeringCover: Bool = true
+    @Published var usesTextHeightForLabels: Bool = false
+
+    init(style: MSFShimmerStyle) {
+        self.style = style
+        super.init()
+    }
+
+    convenience init(style: MSFShimmerStyle,
+                     shouldAddShimmeringCover: Bool = true,
+                     usesTextHeightForLabels: Bool = false) {
+        self.init(style: style)
+        self.shouldAddShimmeringCover = shouldAddShimmeringCover
+        self.usesTextHeightForLabels = usesTextHeightForLabels
+    }
 }
